rename to `fg_aware_rule_based` algorithm
